Introduce module tests for GTFS snapshot
